feat: introduce constants enum
- Introduce a constants enum to define model and API related constants.
- This improves code readability and maintainability.
